requirements.txt
httpx
openai
pyzbar
opencv-python
numpy
Faker
qrcode[pil]
Pillow
tqdm
